updated dependencies, moved sauce demo pages to separate page object class (#111)
