+++ test-suite/testng.xml
@@ -3,7 +3,6 @@
 <suite name="Selenium 4 POC Tests ">
     <suite-files>
         <suite-file path="testng-saucedemo.xml"/>
-        <suite-file path="testng-automationpractice.xml"/>
         <suite-file path="testng-theinternet.xml"/>
         <suite-file path="testng-juice-shop.xml"/>
         <suite-file path="testng-lambdatestecommerce.xml"/>
